EY-3813: Auto-lagring for oppsummeringsteget. Bruker toast også for å vise feil, for det kan hende brukeren ikke ser den som vises ved siden av lagre-knappen.
